+++ src/js/core/navbar.js
@@ -56,7 +56,7 @@ export default function (UIkit) {
 
             if (this.dropbar) {
                 this.dropbar = query(this.dropbar, this.$el) || $('<div></div>').insertAfter(this.dropbarAnchor || this.$el);
-                UIkit.navbarDropbar(this.dropbar, {mode: this.dropbarMode, duration: this.duration, navbar: this});
+                UIkit.navbarDropbar(this.dropbar, {clsDrop: this.clsDrop, mode: this.dropbarMode, duration: this.duration, navbar: this});
             }
 
         },
@@ -102,6 +102,7 @@ export default function (UIkit) {
         mixins: [Class],
 
         defaults: {
+            clsDrop: '',
             mode: 'slide',
             navbar: null,
             duration: 200
@@ -127,7 +128,7 @@ export default function (UIkit) {
             },
 
             beforeshow(e, {$el}) {
-                $el.addClass(`${this.clsDrop}-dropbar`);
+                this.clsDrop && $el.addClass(`${this.clsDrop}-dropbar`);
                 this.transitionTo($el.outerHeight(true));
             },